Extract all html5lib things into a shim module
This creates an intermediary html5lib_shim module between the html5lib
library and Bleach. This will make it easier to update html5lib and fix
any issues that arise from that.

We're going to treat this as an incompatible API change, so changing
the upcoming version to 3.0.0.

Fixes #316
