+++ bleach/__init__.py
@@ -20,7 +20,7 @@ from bleach.sanitizer import (
 # yyyymmdd
 __releasedate__ = ''
 # x.y.z or x.y.z.dev0 -- semver
-__version__ = '2.1.5.dev0'
+__version__ = '3.0.0.dev0'
 VERSION = parse_version(__version__)
