+++ bleach/sanitizer.py
@@ -1,34 +1,15 @@
 from __future__ import unicode_literals
+
 from itertools import chain
 import re
-import string
 
 import six
 from six.moves.urllib.parse import urlparse
 from xml.sax.saxutils import unescape
 
-from bleach._vendor import html5lib
-from bleach._vendor.html5lib.constants import (
-    entities,
-    namespaces,
-    prefixes,
-    tokenTypes,
-)
-from bleach._vendor.html5lib.constants import _ReparseException as ReparseException
-from bleach._vendor.html5lib.filters.base import Filter
-from bleach._vendor.html5lib.filters import sanitizer
-from bleach._vendor.html5lib.serializer import HTMLSerializer
-from bleach._vendor.html5lib._tokenizer import HTMLTokenizer
-from bleach._vendor.html5lib._trie import Trie
-
+from bleach import html5lib_shim
 from bleach.utils import alphabetize_attributes, force_unicode
 
-
-#: Map of entity name to expanded entity
-ENTITIES = entities
-
-#: Trie of html entity string -> character representation
-ENTITIES_TRIE = Trie(ENTITIES)
 
 #: List of allowed tags
 ALLOWED_TAGS = [
@@ -54,16 +35,11 @@ ALLOWED_ATTRIBUTES = {
     'acronym': ['title'],
 }
 
-
 #: List of allowed styles
 ALLOWED_STYLES = []
 
-
 #: List of allowed protocols
 ALLOWED_PROTOCOLS = ['http', 'https', 'mailto']
-
-
-AMP_SPLIT_RE = re.compile('(&)')
 
 #: Invisible characters--0 to and including 31 except 9 (tab), 10 (lf), and 13 (cr)
 INVISIBLE_CHARACTERS = ''.join([chr(c) for c in chain(range(0, 9), range(11, 13), range(14, 32))])
@@ -77,90 +53,6 @@ INVISIBLE_CHARACTERS_RE = re.compile(
 #: String to replace invisible characters with. This can be a character, a
 #: string, or even a function that takes a Python re matchobj
 INVISIBLE_REPLACEMENT_CHAR = '?'
-
-
-def convert_entity(value):
-    """Convert an entity (minus the & and ; part) into what it represents
-
-    This handles numeric, hex, and text entities.
-
-    :arg value: the string (minus the ``&`` and ``;`` part) to convert
-
-    :returns: unicode character or None if it's an ambiguous ampersand that
-        doesn't match a character entity
-
-    """
-    if value[0] == '#':
-        if value[1] in ('x', 'X'):
-            return six.unichr(int(value[2:], 16))
-        return six.unichr(int(value[1:], 10))
-
-    return ENTITIES.get(value, None)
-
-
-def convert_entities(text):
-    """Converts all found entities in the text
-
-    :arg text: the text to convert entities in
-
-    :returns: unicode text with converted entities
-
-    """
-    if '&' not in text:
-        return text
-
-    new_text = []
-    for part in next_possible_entity(text):
-        if not part:
-            continue
-
-        if part.startswith('&'):
-            entity = match_entity(part)
-            if entity is not None:
-                converted = convert_entity(entity)
-
-                # If it's not an ambiguous ampersand, then replace with the
-                # unicode character. Otherwise, we leave the entity in.
-                if converted is not None:
-                    new_text.append(converted)
-                    remainder = part[len(entity) + 2:]
-                    if part:
-                        new_text.append(remainder)
-                    continue
-
-        new_text.append(part)
-
-    return u''.join(new_text)
-
-
-class BleachHTMLTokenizer(HTMLTokenizer):
-    def consumeEntity(self, allowedChar=None, fromAttribute=False):
-        # We don't want to consume and convert entities, so this overrides the
-        # html5lib tokenizer's consumeEntity so that it's now a no-op.
-        #
-        # However, when that gets called, it's consumed an &, so we put that in
-        # the stream.
-        if fromAttribute:
-            self.currentToken['data'][-1][1] += '&'
-
-        else:
-            self.tokenQueue.append({"type": tokenTypes['Characters'], "data": '&'})
-
-
-class BleachHTMLParser(html5lib.HTMLParser):
-    def _parse(self, stream, innerHTML=False, container="div", scripting=False, **kwargs):
-        # Override HTMLParser so we can swap out the tokenizer for our own.
-        self.innerHTMLMode = innerHTML
-        self.container = container
-        self.scripting = scripting
-        self.tokenizer = BleachHTMLTokenizer(stream, parser=self, **kwargs)
-        self.reset()
-
-        try:
-            self.mainLoop()
-        except ReparseException:
-            self.reset()
-            self.mainLoop()
 
 
 class Cleaner(object):
@@ -223,9 +115,9 @@ class Cleaner(object):
         self.strip_comments = strip_comments
         self.filters = filters or []
 
-        self.parser = BleachHTMLParser(namespaceHTMLElements=False)
-        self.walker = html5lib.getTreeWalker('etree')
-        self.serializer = BleachHTMLSerializer(
+        self.parser = html5lib_shim.BleachHTMLParser(namespaceHTMLElements=False)
+        self.walker = html5lib_shim.getTreeWalker('etree')
+        self.serializer = html5lib_shim.BleachHTMLSerializer(
             quote_attr_values='always',
             omit_optional_tags=False,
             escape_lt_in_attrs=True,
@@ -325,80 +217,7 @@ def attribute_filter_factory(attributes):
     raise ValueError('attributes needs to be a callable, a list or a dict')
 
 
-def match_entity(stream):
-    """Returns first entity in stream or None if no entity exists
-
-    Note: For Bleach purposes, entities must start with a "&" and end with
-    a ";".
-
-    :arg stream: the character stream
-
-    :returns: ``None`` or the entity string without "&" or ";"
-
-    """
-    # Nix the & at the beginning
-    if stream[0] != '&':
-        raise ValueError('Stream should begin with "&"')
-
-    stream = stream[1:]
-
-    stream = list(stream)
-    possible_entity = ''
-    end_characters = '<&=;' + string.whitespace
-
-    # Handle number entities
-    if stream and stream[0] == '#':
-        possible_entity = '#'
-        stream.pop(0)
-
-        if stream and stream[0] in ('x', 'X'):
-            allowed = '0123456789abcdefABCDEF'
-            possible_entity += stream.pop(0)
-        else:
-            allowed = '0123456789'
-
-        # FIXME(willkg): Do we want to make sure these are valid number
-        # entities? This doesn't do that currently.
-        while stream and stream[0] not in end_characters:
-            c = stream.pop(0)
-            if c not in allowed:
-                break
-            possible_entity += c
-
-        if possible_entity and stream and stream[0] == ';':
-            return possible_entity
-        return None
-
-    # Handle character entities
-    while stream and stream[0] not in end_characters:
-        c = stream.pop(0)
-        if not ENTITIES_TRIE.has_keys_with_prefix(possible_entity):
-            break
-        possible_entity += c
-
-    if possible_entity and stream and stream[0] == ';':
-        return possible_entity
-
-    return None
-
-
-def next_possible_entity(text):
-    """Takes a text and generates a list of possible entities
-
-    :arg text: the text to look at
-
-    :returns: generator where each part (except the first) starts with an
-        "&"
-
-    """
-    for i, part in enumerate(AMP_SPLIT_RE.split(text)):
-        if i == 0:
-            yield part
-        elif i % 2 == 0:
-            yield '&' + part
-
-
-class BleachSanitizerFilter(sanitizer.Filter):
+class BleachSanitizerFilter(html5lib_shim.SanitizerFilter):
     """html5lib Filter that sanitizes text
 
     This filter can be used anywhere html5lib filters can be used.
@@ -430,14 +249,13 @@ class BleachSanitizerFilter(sanitizer.Filter):
 
         """
         self.attr_filter = attribute_filter_factory(attributes)
-
         self.strip_disallowed_elements = strip_disallowed_elements
         self.strip_html_comments = strip_html_comments
 
         return super(BleachSanitizerFilter, self).__init__(source, **kwargs)
 
     def __iter__(self):
-        for token in Filter.__iter__(self):
+        for token in html5lib_shim.Filter.__iter__(self):
             ret = self.sanitize_token(token)
 
             if not ret:
@@ -523,12 +341,12 @@ class BleachSanitizerFilter(sanitizer.Filter):
 
         # For each possible entity that starts with a "&", we try to extract an
         # actual entity and re-tokenize accordingly
-        for part in next_possible_entity(data):
+        for part in html5lib_shim.next_possible_entity(data):
             if not part:
                 continue
 
             if part.startswith('&'):
-                entity = match_entity(part)
+                entity = html5lib_shim.match_entity(part)
                 if entity is not None:
                     new_tokens.append({'type': 'Entity', 'name': entity})
                     # Length of the entity plus 2--one for & at the beginning
@@ -556,7 +374,7 @@ class BleachSanitizerFilter(sanitizer.Filter):
         # different than the original value.
 
         # Convert all character entities in the value
-        new_value = convert_entities(value)
+        new_value = html5lib_shim.convert_entities(value)
 
         # Nix backtick, space characters, and control characters
         new_value = re.sub(
@@ -645,7 +463,9 @@ class BleachSanitizerFilter(sanitizer.Filter):
 
                 # Drop href and xlink:href attr for svg elements with non-local IRIs
                 if (None, token['name']) in self.svg_allow_local_href:
-                    if namespaced_name in [(None, 'href'), (namespaces['xlink'], 'href')]:
+                    if namespaced_name in [
+                            (None, 'href'), (html5lib_shim.namespaces['xlink'], 'href')
+                    ]:
                         if re.search(r'^\s*[^#\s]', val):
                             continue
 
@@ -676,10 +496,10 @@ class BleachSanitizerFilter(sanitizer.Filter):
 
                 # Figure out namespaced name if the namespace is appropriate
                 # and exists; if the ns isn't in prefixes, then drop it.
-                if ns is None or ns not in prefixes:
+                if ns is None or ns not in html5lib_shim.prefixes:
                     namespaced_name = name
                 else:
-                    namespaced_name = '%s:%s' % (prefixes[ns], name)
+                    namespaced_name = '%s:%s' % (html5lib_shim.prefixes[ns], name)
 
                 attrs.append(' %s="%s"' % (
                     namespaced_name,
@@ -704,7 +524,7 @@ class BleachSanitizerFilter(sanitizer.Filter):
     def sanitize_css(self, style):
         """Sanitizes css in style tags"""
         # Convert entities in the style so that it can be parsed as CSS
-        style = convert_entities(style)
+        style = html5lib_shim.convert_entities(style)
 
         # Drop any url values before we do anything else
         style = re.compile('url\s*\(\s*[^\s)]+?\s*\)\s*').sub(' ', style)
@@ -737,59 +557,3 @@ class BleachSanitizerFilter(sanitizer.Filter):
                 clean.append(prop + ': ' + value + ';')
 
         return ' '.join(clean)
-
-
-class BleachHTMLSerializer(HTMLSerializer):
-    """Wraps the HTMLSerializer and undoes & -> &amp; in attributes"""
-    def escape_base_amp(self, stoken):
-        """Escapes bare & in HTML attribute values"""
-        # First, undo what the HTMLSerializer did
-        stoken = stoken.replace('&amp;', '&')
-
-        # Then, escape any bare &
-        for part in next_possible_entity(stoken):
-            if not part:
-                continue
-
-            if part.startswith('&'):
-                entity = match_entity(part)
-                # Only leave entities in that are not ambiguous. If they're
-                # ambiguous, then we escape the ampersand.
-                if entity is not None and convert_entity(entity) is not None:
-                    yield '&' + entity + ';'
-
-                    # Length of the entity plus 2--one for & at the beginning
-                    # and and one for ; at the end
-                    part = part[len(entity) + 2:]
-                    if part:
-                        yield part
-                    continue
-
-            yield part.replace('&', '&amp;')
-
-    def serialize(self, treewalker, encoding=None):
-        """Wrap HTMLSerializer.serialize and escape bare & in attributes"""
-        in_tag = False
-        after_equals = False
-
-        for stoken in super(BleachHTMLSerializer, self).serialize(treewalker, encoding):
-            if in_tag:
-                if stoken == '>':
-                    in_tag = False
-
-                elif after_equals:
-                    if stoken != '"':
-                        for part in self.escape_base_amp(stoken):
-                            yield part
-
-                        after_equals = False
-                        continue
-
-                elif stoken == '=':
-                    after_equals = True
-
-                yield stoken
-            else:
-                if stoken.startswith('<'):
-                    in_tag = True
-                yield stoken
